Begin the creation of a basic daemon for PlayingWHAT
The purpose of this daemon is to provide a way for PlayingWHAT to
communicate with the Spotify API at a periodic interval so it knows what
the user is currently playing and update the InkyWHAT display with the
track that the user is playing (or has played previously).

For now, this commit introduces the following for the daemon:

- The ability to start a daemon (via forking)
- The ability to stop the daemon
- The ability to write a PID file to /run/playwhat so that the client
  controller that stops the daemons know what PID the daemon is running
  under
- The ability to create a Unix socket at /run/playwhat so that the
  client controller can communicate with the daemon (where Unix signals
  wouldn't be enough)

Note that the current implementation expects /run/playwhat to be created
and writeable to the service. Note that this daemon is best integrated
with systemd (a sample systemd file will be created in a future commit).