Add the ability to update the display
The playwhat.daemon module (when run using python3 -m playwhat.daemon)
now supports updating what'll be shown on the display. This should be
used for debugging purposes.

In addition, provide a MessageHandler (and StreamMesasgeHandler) that'll
be used by the daemon client and server for handling messages that both
parts of the daemon receive from each other (i.e., requests from clients
and responses from server).