Show the "like" status of a track on the InkyWHAT
Add a heart icon that'll show if the user has liked the track that's
currently played. If the track hasn't been liked, than an outline of a
heart will be shown instead.

In addition, change PainterOptions to take in a is_liked Boolean
property, which is used by the painter to determine which heart icon it
should show on the InkyWHAT.