Fix an issue where the track name could be truncated
During the part of playwhat.painter.utils.wrap_and_ellipsize_text when
we check to see if we're about to exceed the maximum lines of text we
can wrap, it's possible when assembling the remaining current_line_str
that the last word in the track will get truncated (or a single word may
be missing) (curse you off by one errors! (╯‵□′)╯︵┻━┻)

Fix this issue by not doing "index + 1" when assembling the remaining
current_line_str, before sending that over to the ellipsize_text
function.