Align middle content to middle of the Inky wHAT screen
If CONTENT_START_Y is None, then align the middle content (i.e., album
art, track title, album name, artist, and duration) in the middle of the
Inky wHAT screen (using the dimension of the album art to determine the
height of the middle content).

This should make the middle content appear in the center better, now
that we've adjusted the padding of the PlayingWHAT screen from 20 pixels
to 10 pixels in our earlier commit.