Don't return the Spotipy client if we can't access cached credentials
As we can't get the refresh token to use to refresh our access to
Spotify's API, this should be considered an error (and so, the service
should try polling again in 60 seconds rather than erroring out and than
polling again).

In addition, show an informational message telling the user that they
will need to re-authenticate with Spotify's API (by running the
`python3 -m playwhat.console -m auth` command again).