Use python-dateutil package for creating time-zone aware datetimes
The python-dateutil package provides some additional helpful methods for
working with Python's datetime module.

In particular, PlayingWHAT will use python-dateutil so it can create a
UTC-aware datetime for storing when a track was played (as the Spotify
API returns a UTC timestamp).

This should fix some issues where the "Recently Played" screen was
showing time differences that were off due to using datetime that wasn't
UTC-aware.