Account for top and bottom padding when calculating content_y
Ensure that we consider the top and bottom padding (in addition to the
height of the main content and the image height) when figuring out how
to center our content.

This should fix an issue where the main content of the "Now Playing"
and "Recently Played" looks off-centered (especially prominent with the
"Recently Played" screen and the list of tracks shown).