Allow the number of recent tracks to be shown to be adjustable
Use the SHOW_RECENT_TRACKS environment variable and change it so that
the user can enter an integer, which will indicate the number of recent
tracks to show on the Inky wHAT screen.

However, the user can disable that and show the "Not Playing" screen by
setting this environment variable to 0. :slight_smile: