Fix word-wrapping for single (yet long) words
When using the wrap_and_ellipsize_text method in playwhat.painter.utils
module, if the text being wrap and ellipsized is only a single word
("single word" as in we don't notice any spaces in the text), then use
ellipsize_text (inside playwhat.painter.utils) to determine how to
ellipsize the long word.

This should fix an issue when Inky wHAT will add a line break for a long
word for a track even though that line break is not needed, e.g. for the
track "ヒャダインのカカカタ☆カタオモイ-C" by HYADAIN (the opening track
for Nichijou), a line break was inserted before the track name.