Draw the "Times Played" part separately from the track's name
This is so that we can ellipsize the track's name (but still show the
"Times Played" part), since the "Times Played" part can become
ellipsized (and therefore, not shown) if the track's name is long.

In addition, show the "Times Played" part in red so that it'll stand
out.

Finally, only show the "Times Played" if the song has been played more
than once consecutively. (We don't want to be showing (× 1) for all
tracks. :sweat_smile:)