Support pagination when building the Recently Played screen
If the user (say me) has a song that is put on repeat multiple time,
it's possible that the previous iteration (of PlayingWHAT's "Recently
Played" screen) will not fill the table up to whatever the maximum
number of tracks that should be shown on the screen, because the user
happened to play a track 50 times, so the screen will show that track (×
50), for a total of a single track instead of seven tracks (the default
number of recently-played tracks that'll show on that screen).

Rather than calling quits early, try to page through the
current_user_recently_played() results until there are no more results
available and then call quit (in the off chance that although the user
played a single track 50 times, there happened to be six other songs
that were played previously that could be shown on the "Recently Played"
screen. :sweat_smile:)