Begin the rewrite process of PlayingWHAT
PlayingWHAT will be rewritten to be bigger and better than before! And
by bigger and better, I'm hoping to make PlayingWHAT work _primarily_ on
the Inky Impression screen (no guarantee it'll work on the Inky wHAT
like before, but that may get reinstated in some future commits later,
once the Inky Impression build is complete) and make it easier to manage
(via the user's web browser instead of having to the use the
playwhat.console application). :smile: